+++ back/src/bsdasris/validation.ts
@@ -33,7 +33,7 @@ import {
   WeightUnits,
   transporterRecepisseSchema
 } from "../common/validation";
-import { getOperationModesFromOperationCode } from "../common/operationModes";
+import { destinationOperationModeValidation } from "../common/validation/operationMode";
 
 const wasteCodes = DASRI_WASTE_CODES.map(el => el.code);
 
@@ -650,43 +650,7 @@ export const operationSchema: FactorySchemaOf<
           return true;
         }
       ),
-    destinationOperationMode: yup
-      .mixed<OperationMode | null | undefined>()
-      .oneOf([...Object.values(OperationMode), null, undefined])
-      .nullable()
-      .test(
-        "processing-mode-matches-processing-operation",
-        "Le mode de traitement n'est pas compatible avec l'opération de traitement choisie",
-        function (item) {
-          if (!context.operationSignature) return true;
-
-          const { destinationOperationCode } = this.parent;
-          const destinationOperationMode = item;
-
-          if (destinationOperationCode) {
-            const modes = getOperationModesFromOperationCode(
-              destinationOperationCode
-            );
-
-            if (modes.length && !destinationOperationMode) {
-              return new yup.ValidationError(
-                "Vous devez préciser un mode de traitement"
-              );
-            } else if (
-              (modes.length &&
-                destinationOperationMode &&
-                !modes.includes(destinationOperationMode)) ||
-              (!modes.length && destinationOperationMode)
-            ) {
-              return new yup.ValidationError(
-                "Le mode de traitement n'est pas compatible avec l'opération de traitement choisie"
-              );
-            }
-          }
-
-          return true;
-        }
-      ),
+    destinationOperationMode: destinationOperationModeValidation(context),
     destinationOperationDate: yup
       .date()
       .label("Date de traitement")
